requirements.txt
streamlit
pandas
numpy
openai
faiss-cpu
scikit-learn
litellm
rank-bm25
loguru
requests
pyperclip
google-cloud-bigquery
pyarrow
pytz
streamlit-extras
ollama